script: Limit max extern args to 16 bit

+++ libs/script/include/script_args.h
@@ -12,16 +12,16 @@ typedef struct sScriptEnum ScriptEnum;
 
 typedef struct {
   const ScriptVal* values;
-  usize            count;
+  u16              count;
 } ScriptArgs;
 
-f64          script_arg_maybe_number(ScriptArgs, u32 i, f64 def);
-bool         script_arg_maybe_bool(ScriptArgs, u32 i, bool def);
-GeoVector    script_arg_maybe_vector3(ScriptArgs, u32 i, GeoVector def);
-GeoQuat      script_arg_maybe_quat(ScriptArgs, u32 i, GeoQuat def);
-EcsEntityId  script_arg_maybe_entity(ScriptArgs, u32 i, EcsEntityId def);
-StringHash   script_arg_maybe_string(ScriptArgs, u32 i, StringHash def);
-TimeDuration script_arg_maybe_time(ScriptArgs, u32 i, TimeDuration def);
-i32          script_arg_maybe_enum(ScriptArgs, u32 i, const ScriptEnum*, i32 def);
+f64          script_arg_maybe_number(ScriptArgs, u16 i, f64 def);
+bool         script_arg_maybe_bool(ScriptArgs, u16 i, bool def);
+GeoVector    script_arg_maybe_vector3(ScriptArgs, u16 i, GeoVector def);
+GeoQuat      script_arg_maybe_quat(ScriptArgs, u16 i, GeoQuat def);
+EcsEntityId  script_arg_maybe_entity(ScriptArgs, u16 i, EcsEntityId def);
+StringHash   script_arg_maybe_string(ScriptArgs, u16 i, StringHash def);
+TimeDuration script_arg_maybe_time(ScriptArgs, u16 i, TimeDuration def);
+i32          script_arg_maybe_enum(ScriptArgs, u16 i, const ScriptEnum*, i32 def);
 
 ScriptVal script_arg_last_or_null(ScriptArgs);
