+++ libs/script/src/args.c
@@ -2,35 +2,35 @@
 #include "script_enum.h"
 #include "script_val.h"
 
-f64 script_arg_maybe_number(const ScriptArgs args, const u32 i, const f64 def) {
+f64 script_arg_maybe_number(const ScriptArgs args, const u16 i, const f64 def) {
   return args.count > i ? script_get_number(args.values[i], def) : def;
 }
 
-bool script_arg_maybe_bool(const ScriptArgs args, const u32 i, const bool def) {
+bool script_arg_maybe_bool(const ScriptArgs args, const u16 i, const bool def) {
   return args.count > i ? script_get_bool(args.values[i], def) : def;
 }
 
-GeoVector script_arg_maybe_vector3(const ScriptArgs args, const u32 i, const GeoVector def) {
+GeoVector script_arg_maybe_vector3(const ScriptArgs args, const u16 i, const GeoVector def) {
   return args.count > i ? script_get_vector3(args.values[i], def) : def;
 }
 
-GeoQuat script_arg_maybe_quat(const ScriptArgs args, const u32 i, const GeoQuat def) {
+GeoQuat script_arg_maybe_quat(const ScriptArgs args, const u16 i, const GeoQuat def) {
   return args.count > i ? script_get_quat(args.values[i], def) : def;
 }
 
-EcsEntityId script_arg_maybe_entity(const ScriptArgs args, const u32 i, const EcsEntityId def) {
+EcsEntityId script_arg_maybe_entity(const ScriptArgs args, const u16 i, const EcsEntityId def) {
   return args.count > i ? script_get_entity(args.values[i], def) : def;
 }
 
-StringHash script_arg_maybe_string(const ScriptArgs args, const u32 i, const StringHash def) {
+StringHash script_arg_maybe_string(const ScriptArgs args, const u16 i, const StringHash def) {
   return args.count > i ? script_get_string(args.values[i], def) : def;
 }
 
-TimeDuration script_arg_maybe_time(const ScriptArgs args, const u32 i, const TimeDuration def) {
+TimeDuration script_arg_maybe_time(const ScriptArgs args, const u16 i, const TimeDuration def) {
   return args.count > i ? script_get_time(args.values[i], def) : def;
 }
 
-i32 script_arg_maybe_enum(const ScriptArgs args, const u32 i, const ScriptEnum* e, const i32 def) {
+i32 script_arg_maybe_enum(const ScriptArgs args, const u16 i, const ScriptEnum* e, const i32 def) {
   if (args.count <= i) {
     return def;
   }
